Moved code to MyProfileViewController
- I was having issues getting a black screen when running the project, so I took all that code I added and pasted it on a new file I created called MyProfileViewController
