+++ labs-ios-starter/Login/ProfileTabBarViewController.swift
@@ -10,28 +10,14 @@ import UIKit
 
 class ProfileTabBarViewController: UITabBarController {
     
-    @IBOutlet weak var profileImageOutlet: UIImageView!
-    
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        //ProfileView
-        imageViewUpdate()
         
         NotificationCenter.default.addObserver(forName: .oktaAuthenticationExpired,
                                                              object: nil,
                                                              queue: .main,
                                                              using: dismissToLogin)
-    }
-    
-    func imageViewUpdate() {
-        guard let myProfileView = profileImageOutlet else { return }
-        myProfileView.layer.borderWidth = 1.0
-        myProfileView.layer.masksToBounds = false
-        myProfileView.layer.borderColor = UIColor.white.cgColor
-        myProfileView.layer.cornerRadius = profileImageOutlet.frame.size.width / 2
-        myProfileView.clipsToBounds = true
-        
     }
     
     func dismissToLogin(_ notification: Notification)  {
